// src/app/@api/model/newsDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AttachmentDto } from './attachmentDto';


export interface NewsDto { 
    content: string;
    image?: AttachmentDto;
    lastUpdatedDate?: Date;
    newsId?: number;
    status?: NewsDto.StatusEnum;
    title: string;
}
export namespace NewsDto {
    export type StatusEnum = 'PUBLISHED' | 'UNPUBLISHED';
    export const StatusEnum = {
        PUBLISHED: 'PUBLISHED' as StatusEnum,
        UNPUBLISHED: 'UNPUBLISHED' as StatusEnum
    };
}